# media/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/transactions/": {
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Create a new Transaction.",
        "operationId": "create_transaction_transactions__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWithTags"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "loc": [
                      "body",
                      "string"
                    ],
                    "msg": "string",
                    "type": "Type[]"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "List Transactions.",
        "operationId": "list_transactions_transactions__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "description",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  },
                  "title": "Response List Transactions Transactions  Get"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "loc": [
                      "body",
                      "string"
                    ],
                    "msg": "string",
                    "type": "Type[]"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get Transaction by ID",
        "operationId": "get_transaction_transactions__transaction_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "loc": [
                      "body",
                      "string"
                    ],
                    "msg": "string",
                    "type": "Type[]"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Transactions"
        ],
        "summary": "Update Transaction",
        "operationId": "update_transaction_transactions__transaction_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Transaction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Transaction"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "loc": [
                      "body",
                      "string"
                    ],
                    "msg": "string",
                    "type": "Type[]"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transactions"
        ],
        "summary": "Delete Transaction",
        "operationId": "delete_transaction_transactions__transaction_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully deleted",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "message": "string"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "loc": [
                      "body",
                      "string"
                    ],
                    "msg": "string",
                    "type": "Type[]"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}/with-category": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get a specific transaction with Category details.",
        "operationId": "get_transaction_with_category_transactions__transaction_id__with_category_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWithCategory"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "loc": [
                      "body",
                      "string"
                    ],
                    "msg": "string",
                    "type": "Type[]"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          }
        }
      }
    },
    "/transactions/{transaction_id}/with-tags": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "Get a specific transaction with Tags details.",
        "operationId": "get_transaction_with_tags_transactions__transaction_id__with_tags_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionWithTags"
                }
              }
            }
          },
          "404": {
            "description": "Entity not found",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "example": {
                  "detail": {
                    "loc": [
                      "body",
                      "string"
                    ],
                    "msg": "string",
                    "type": "Type[]"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization error",
            "content": {
              "application/json": {
                "example": {
                  "detail": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Category": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "$ref": "#/components/schemas/CategoryType"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "type",
          "user_id"
        ],
        "title": "Category"
      },
...
      "CategoryType": {
        "type": "string",
        "enum": [
          "income",
          "expense"
        ],
        "title": "CategoryType"
      },
...
      "Tag": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "user_id"
        ],
        "title": "Tag"
      },
...
      "Transaction": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "category_id",
          "amount",
          "date"
        ],
        "title": "Transaction"
      },
      "TransactionCreate": {
        "properties": {
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Amount"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "tag_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Ids",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "category_id",
          "amount",
          "date"
        ],
        "title": "TransactionCreate"
      },
      "TransactionUpdate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "date",
          "description"
        ],
        "title": "TransactionUpdate"
      },
      "TransactionWithCategory": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "category_id",
          "amount",
          "date",
          "category"
        ],
        "title": "TransactionWithCategory"
      },
      "TransactionWithTags": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id"
          },
          "amount": {
            "type": "string",
            "title": "Amount"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "title": "Date"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "category_id",
          "amount",
          "date",
          "tags"
        ],
        "title": "TransactionWithTags"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}